\docType{data}
\name{n_iris}
\alias{n_iris}
\title{Anderson's Iris Data 1935}
\format{numeric vector of length 3, with all values 49}
\source{
  Anderson 1935
}
\description{
  Measurements of sepal length, sepal width, petal length
  and petal width.  Three groups of 49 Versicolor, 49
  Virginica and 49 Setosa.
}
\author{
  Anderson
}
\keyword{datasets}

