\docType{data}
\name{n_vole}
\alias{n_vole}
\title{sample data from Airoldi and Flury 1988}
\format{numeric vector of length 4}
\source{
  Airoldi and Hoffmann 1988
}
\description{
  This contains 4 sample sizes: 172 M.californicus males,
  140 M. californicus females, 87 M. ochrogaster males and
  75 M. ochrogaster females.
}
\author{
  Airoldi and Hoffmann
}
\keyword{datasets}

