\name{random_rotation}
\alias{random_rotation}
\title{Random Rotation}
\usage{
  random_rotation(M)
}
\arguments{
  \item{M}{matrix to be rotated}
}
\description{
  Applies a random rotation to a matrix based on the
  subgroup algorithm provided by Diaconis, P. and
  Shahshahani, M. (1987).
}
\references{
  Diaconis, P.; Shahshahani, M.  The subgroup algorithm for
  generating uniform random variables.  Probability in the
  Engineering and Informational Sciences.  1987, 1, 15-32.
}

