\name{random_rotation_matrix}
\alias{random_rotation_matrix}
\title{Random Rotation Matrix}
\usage{
  random_rotation_matrix(n)
}
\arguments{
  \item{n}{dimension of matrix}
}
\description{
  Generates a random rotation matrix of dimension n, based
  on the subgroup algorithm provided by Diaconis, P. and
  Shahshahani, M. (1987).
}
\references{
  Diaconis, P.; Shahshahani, M.  The subgroup algorithm for
  generating uniform random variables.  Probability in the
  Engineering and Informational Sciences.  1987, 1, 15-32.
}

