\docType{data}
\name{test_bank}
\alias{test_bank}
\title{Sample Data from FLury and Riedwyl 1988}
\format{a numeric 6x6 matrix}
\source{
  Flury
}
\description{
  Data measuring former Swiss 1000-franc bank notes (widths
  from the left, right,bottom and top sides)
}
\author{
  Flury
}
\keyword{datasets}

