\docType{data}
\name{test_iris}
\alias{test_iris}
\title{Anderson's Iris Data 1935}
\format{a set of three 4x4 matrices}
\source{
  Anderson 1935
}
\description{
  Measurements of sepal length, sepal width, petal length
  and petal width.  Three groups of 49 Versicolor, 49
  Virginica and 49 Setosa.
}
\author{
  Anderson
}
\keyword{datasets}

