\docType{data}
\name{test_marten}
\alias{test_marten}
\title{test_marten}
\format{a set of 2 4x4 matrices}
\source{
  Flury
}
\description{
  Data from Jolicoeur of North American Marten.
  Measurements of Humerus length, Humerus width, Femur
  length and Femur width.
}
\author{
  Flury
}
\keyword{datasets}

