\docType{data}
\name{test_vole}
\alias{test_vole}
\title{Sample data from Airoldi and Hoffmann 1984}
\format{a set of four 3x3 matrices}
\source{
  Airoldi and Hoffmann 1984
}
\description{
  Skull measurements (length, width, height) of small
  rodents. Covariance matrices for 4 groups: 172
  M.californicus males, 140 M. californicus females, 87 M.
  ochrogaster males and 75 M. ochrogaster females.
}
\author{
  Airoldi and Hoffmann
}
\keyword{datasets}

